Updated rendering logic
